server/requirements.txt


Flask==3.0.0
flask-cors==4.0.0
gunicorn==20.1.0
boto3==1.28.0
python-dotenv==1.0.0
pytest==7.4.0
